refactored the SessionView to make room for map component
