+++ client/src/SessionView/index.js
@@ -1,46 +1,61 @@
-import { useState, useCallback } from 'react'
+import { useState } from 'react'
 import { Helmet } from 'react-helmet'
 import {
-  Avatar,
-  Card,
-  CardActions,
-  CardHeader,
-  CardMedia,
-  Dialog,
   Grid,
-  IconButton,
-  Tooltip,
+  Tab,
+  Tabs,
 } from '@material-ui/core'
-import { LocalLibrary, PhotoLibrary } from '@material-ui/icons'
-import { makeStyles } from '@material-ui/core/styles'
+import { Assistant, Map } from '@material-ui/icons'
+import { makeStyles, withStyles } from '@material-ui/core/styles'
 
 import * as factions from '../gameInfo/factions'
+import { getFactionCheatSheetPath } from '../gameInfo/factions'
 
 import useRealTimeSession from './useRealTimeSession'
+import Overview from './Overview'
 import ShuffleFactionsButton from './ShuffleFactionsButton'
 import ShareButton from './ShareButton'
-import VictoryPoints from './VictoryPoints'
-import PublicObjectives from './PublicObjectives'
 
-const useStyles = makeStyles(theme => ({
+const StyledTabs = withStyles({
+  indicator: {
+    display: 'flex',
+    justifyContent: 'center',
+    backgroundColor: 'transparent',
+    '& > span': {
+      maxWidth: 70,
+      width: '100%',
+      backgroundColor: '#fff',
+    },
+  },
+})((props) => <Tabs {...props} TabIndicatorProps={{ children: <span /> }} />);
+
+const StyledTab = withStyles((theme) => ({
   root: {
-    color: 'white',
+    textTransform: 'none',
+    color: '#fff',
+    fontWeight: theme.typography.fontWeightRegular,
+    fontSize: theme.typography.pxToRem(15),
+    minHeight: 'unset',
+    marginRight: theme.spacing(1),
+    '&:focus': {
+      opacity: 1,
+    },
   },
-  factionCard: {
-    backgroundColor: 'rgba(255, 255, 255, 0.1)',
-    color: 'white',
+  wrapper: {
+    flexDirection: 'row',
   },
-  factionCardIcon: {
-    color: 'white',
-  },
-  media: {
-    height: 0,
-    paddingTop: '71.25%',
-    cursor: 'pointer',
-  },
-}))
+}))((props) => <Tab disableRipple {...props} />);
 
-const getFactionCheatSheetPath = factionKey => `/factionCheatsheets/${factionKey.toLowerCase()}.png`
+const useStyles = makeStyles({
+  header: {
+    marginBottom: '2em',
+  },
+})
+
+const VIEW = {
+  overview: 0,
+  map: 1,
+}
 
 function SessionView({
   session,
@@ -51,18 +66,16 @@ function SessionView({
 }) {
   useRealTimeSession(session.id)
   const classes = useStyles()
-  const [factionDialogOpen, setFactionDialogOpen] = useState(false)
-  const [faction, setFaction] = useState(null)
-
-  const updateFactionPointsInSession = useCallback((faction, points) => updateFactionPoints({
-    sessionId: session.id,
-    faction,
-    points
-  }), [session.id, updateFactionPoints])
 
   const sortedPoints = [...session.points]
   sortedPoints.sort((a, b) => b.points - a.points)
   const winningFaction = sortedPoints[0].faction
+
+  const [view, setView] = useState(VIEW.overview);
+
+  const handleChange = (event, newView) => {
+    setView(newView);
+  };
 
   return <>
     <Helmet>
@@ -74,15 +87,12 @@ function SessionView({
       <meta property="og:image" content={`${window.location.origin}${getFactionCheatSheetPath(winningFaction)}`} />
     </Helmet>
 
-    <Grid
-      className={classes.root}
-      container
-      alignItems="center"
-      justifyContent="center"
-      spacing={4}
-    >
+    <Grid container className={classes.header}>
       <Grid item xs={6}>
-        session from: {new Date(session.createdAt).toLocaleDateString()} {new Date(session.createdAt).toLocaleTimeString()}
+        <StyledTabs value={view} onChange={handleChange} aria-label="styled tabs example">
+          <StyledTab icon={<Assistant />} label="Overview" title="Overview" />
+          <StyledTab icon={<Map />} label="Map" title="Map" />
+        </StyledTabs>
       </Grid>
       <Grid item container xs={6} justifyContent="flex-end">
         <ShuffleFactionsButton
@@ -92,70 +102,19 @@ function SessionView({
         />
         <ShareButton id={session.id} />
       </Grid>
-      <Grid item xs={12}>
-        <VictoryPoints
-          onChange={updateFactionPointsInSession}
-          points={session.points}
-          factions={session.factions}
-        />
-      </Grid>
-      <Grid item xs={12}>
-        <PublicObjectives
-          session={session}
-          updateFactionPoints={updateFactionPoints}
-        />
-      </Grid>
-      {session.factions.map(faction => {
-        const factionData = factions.getData(faction)
-
-        return <Grid item xs={12} sm={6} key={factionData.key}>
-          <Card className={classes.factionCard}>
-            <CardHeader
-              avatar={<Avatar alt={factionData.name} src={factionData.image}/>}
-              title={factionData.name}
-            />
-            <CardMedia
-              onClick={() => {
-                setFaction(factionData.key)
-                setFactionDialogOpen(open => !open)
-              }}
-              className={classes.media}
-              title={factionData.name}
-              image={getFactionCheatSheetPath(factionData.key)}
-            />
-            <CardActions disableSpacing>
-              <Tooltip title="go to wiki" placement="top">
-                <IconButton
-                  className={classes.factionCardIcon}
-                  aria-label="go to wiki"
-                  href={`https://twilight-imperium.fandom.com/wiki/${encodeURIComponent(factionData.name)}`}
-                  target="about:blank"
-                >
-                  <LocalLibrary />
-                </IconButton>
-              </Tooltip>
-              <Tooltip title="open original image" placement="top">
-                <IconButton
-                  className={classes.factionCardIcon}
-                  aria-label="open original image"
-                  href={getFactionCheatSheetPath(factionData.key)}
-                  target="about:blank"
-                >
-                  <PhotoLibrary />
-                </IconButton>
-              </Tooltip>
-            </CardActions>
-          </Card>
-        </Grid>
-      })}
     </Grid>
-    <Dialog
-      open={factionDialogOpen}
-      onClose={() => setFactionDialogOpen(false)}
-      maxWidth='lg'
-    >
-      {faction && <img src={getFactionCheatSheetPath(faction)} alt={faction} />}
-    </Dialog>
+
+    <div hidden={view !== VIEW.overview }>
+      <Overview
+        session={session}
+        shuffleFactions={shuffleFactions}
+        setFactions={setFactions}
+        updateFactionPoints={updateFactionPoints}
+      />
+    </div>
+    <div hidden={view !== VIEW.map }>
+      <p>hello map here</p>
+    </div>
   </>
 }
